implemented finding all photos for logged in user, adding photo for logged in user

+++ taggr front end and back end starting code/final-capstone - Copy/java/src/main/java/com/techelevator/dao/UserSqlDAO.java
@@ -23,8 +23,8 @@ public class UserSqlDAO implements UserDAO {
     }
 
     @Override
-    public int findIdByUsername(String username) {
-        return jdbcTemplate.queryForObject("select user_id from users where username = ?", int.class, username);
+    public Long findIdByUsername(String username) {
+        return jdbcTemplate.queryForObject("select user_id from users where username = ?", Long.class, username);
     }
 
 	@Override
